refactor(job): use SDK logger for cost warning; restructure tests
Address review feedback from @LinoGiger:
- Emit the create-time cost warning via the SDK logger (logger.warning,
  the repo's user-facing warning convention) instead of warnings.warn.
- Move tests into a package layout mirroring src
  (tests/rapidata_client/{job,audience}/) instead of one flat file, and
  keep the AudienceJobState-based test helper from the polling refactor.
